usb: host: use uint16_t for Language ID
usbh truncates id when calling request description in function usbh_req_desc. This issue is mostly visible in string descriptor requests, as only then can wIndex be larger than 0xFF. Notably, this affects commonly used language IDs like English (United States) which is 0x0409.

+++ subsys/usb/host/usbh_ch9.h
@@ -22,7 +22,7 @@ int usbh_req_setup(struct usb_device *const udev,
 
 int usbh_req_desc(struct usb_device *const udev,
 		  const uint8_t type, const uint8_t index,
-		  const uint8_t id,
+		  const uint16_t id,
 		  const uint16_t len,
 		  struct net_buf *const data);
